refactor reed solomon code
