+++ primitives/src/lib.rs
@@ -22,12 +22,12 @@ pub mod commitment;
 pub mod constants;
 pub mod crhf;
 pub mod elgamal;
-pub mod erasure_code;
 pub mod errors;
 pub mod hash_to_group;
 pub mod merkle_tree;
 pub mod pcs;
 pub mod prf;
+pub mod reed_solomon_code;
 pub mod rescue;
 pub mod signatures;
 pub mod toeplitz;
